+++ sync-for-commerce/src/codatsynccommerce/models/operations/list_companies.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 import dataclasses
 import requests as requests_http
 from ..shared import companies as shared_companies
-from typing import Optional
+from typing import Callable, Optional
 
 
 
@@ -25,6 +25,7 @@ class ListCompaniesRequest:
 
 @dataclasses.dataclass
 class ListCompaniesResponse:
+    next: Callable[[], Optional[ListCompaniesResponse]] = dataclasses.field()
     content_type: str = dataclasses.field()
     status_code: int = dataclasses.field()
     companies: Optional[shared_companies.Companies] = dataclasses.field(default=None)