+++ sync-for-commerce/src/codatsynccommerce/models/operations/list_connections.py
@@ -5,7 +5,7 @@ import dataclasses
 import requests as requests_http
 from ..shared import connections as shared_connections
 from ..shared import errormessage as shared_errormessage
-from typing import Optional
+from typing import Callable, Optional
 
 
 
@@ -27,6 +27,7 @@ class ListConnectionsRequest:
 
 @dataclasses.dataclass
 class ListConnectionsResponse:
+    next: Callable[[], Optional[ListConnectionsResponse]] = dataclasses.field()
     content_type: str = dataclasses.field()
     status_code: int = dataclasses.field()
     connections: Optional[shared_connections.Connections] = dataclasses.field(default=None)